Refactored to have base sender and receiver classes
